deploy/funcion_oauth_google/requirements.txt
google-auth
google-auth-oauthlib
google-api-python-client
google-cloud-storage
functions-framework